HomeWork_11/pyproject.toml
[tool.poetry]
name = "homework-11"
version = "0.1.0"
description = ""
authors = ["YpL68 <yuriy.plks@gmail.com>"]
readme = "README.md"
packages = [{include = "homework_11"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
flask-migrate = "^4.0.0"
marshmallow = "^3.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
